core: lib: import ChaCha20-Poly1305 from LibTomCrypt
Currently AES-GCM (or AES-CCM) is used to encrypt and integrity protect files
in secure storage.
A newer and widely used AEAD algorithm, ChaCha20-Poly1305, can be
exploited for secure storage since it can perform better than software
implementeation of AES. Moreover it is not susceptible to timing attacks
and it has no practical limit for key/nonce pairs.
LibTomCrypt has already an implementation of such an algorithm but cannot
be used yet in OP-TEE source code.

This commit adds the initial support of ChaCha20-Poly1305 in crytpo lib by
using the implementation in LibTomCrypt, following the AES import approach
and enables it by default.

Signed-off-by: Luca Miccio <lucmiccio@gmail.com>
Reviewed-by: Paolo Valente <paolo.valente@unimore.it>