core, fs_htree: select AEAD algorithm with compile option
Currently choosing the default AEAD algorithm for components in secure
storage is not so trivial because it involves manual modifications
to files (e.g. mostly change all the defines in the Hash tree component).
Since it is now possible to choose between AES-GCM(CCM) and ChaCha20-Poly1305
algorithms, manually changing files is not sustainable anymore because
most of the parameters vary from a method to another. Thus, this
process could be potentially error-prone.

This commit adds a new compile option in order to simplfiy the selection of
the preferred authentication method for Hash tree component.

Since the support of ChaCha20-Poly1305 is experimental, AES-GCM is set as
default.

Signed-off-by: Luca Miccio <lucmiccio@gmail.com>
Reviewed-by: Paolo Valente <paolo.valente@unimore.it>